+++ R/zzz.R
@@ -5,8 +5,8 @@
 #' @importFrom mlr3 mlr_learners LearnerClassif LearnerRegr
 "_PACKAGE"
 
-.onLoad = function(libname, pkgname) {
-  # nocov start
+# nocov start
+register_mlr3 = function(libname, pkgname) {
   # get mlr_learners dictionary from the mlr3 namespace
   x = utils::getFromNamespace("mlr_learners", ns = "mlr3")
 
@@ -15,4 +15,20 @@
   x$add("regr.gamboost", LearnerRegrGAMBoost)
   x$add("classif.glmboost", LearnerClassifGLMBoost)
   x$add("regr.glmboost", LearnerRegrGLMBoost)
-} # nocov end
+}
+
+.onLoad = function(libname, pkgname) {
+  register_mlr3extratrees()
+  setHook(packageEvent("mlr3", "onLoad"), function(...) register_mlr3(),
+    action = "append")
+}
+
+.onUnload = function(libpath) {
+  # nocov start
+  event = packageEvent("mlr3", "onLoad")
+  hooks = getHook(event)
+  pkgname = vapply(hooks, function(x) environment(x)$pkgname, NA_character_)
+  setHook(event, hooks[pkgname != "mlr3learners.extratrees"],
+    action = "replace")
+}
+# nocov end