+++ NEWS.md
@@ -1,7 +1,7 @@
 # mlr3learners.mboost 0.2.0.9000
 
 - Use `$help()` method from super class
-
+- Optimize .onLoad() and add .Unload()
 
 # mlr3learners.mboost 0.2.0